requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.13
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
groq==0.4.1

# Document processing
pymupdf==1.23.7
pdfminer.six==20221105
python-magic==0.4.27

# Vector store
chromadb==0.4.18
faiss-cpu==1.7.4

# Utilities
tqdm==4.66.1
numpy==1.26.2
pandas==2.1.3
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging and monitoring
loguru==0.7.2
